# docs/swagger.yaml
definitions:
...
  api_response.CreateCommentResponse:
    properties:
      comment:
        $ref: '#/definitions/models.Comment'
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  api_response.DeleteCommentResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  api_response.GetCommentResponse:
    properties:
      comment:
        $ref: '#/definitions/models.Comment'
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
  api_response.GetCommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/types.Pagination'
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  api_response.UpdateCommentResponse:
    properties:
      comment:
        $ref: '#/definitions/models.Comment'
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  models.Comment:
    properties:
      comment:
        maxLength: 1000
        type: string
      created_at:
        type: string
      id:
        type: integer
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    required:
    - comment
    type: object
  models.CreateCommentRequest:
    properties:
      comment:
        maxLength: 1000
        type: string
      ticket_id:
        type: integer
    required:
    - comment
    - ticket_id
    type: object
...
  models.Ticket:
    properties:
      assigned_by:
        $ref: '#/definitions/models.User'
      assigned_by_id:
        description: Who assigned the ticket
        type: integer
      assignee:
        $ref: '#/definitions/models.User'
      assignee_id:
        type: integer
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by_id:
        description: Who created the ticket
        type: integer
      createdBy:
        $ref: '#/definitions/models.User'
      description:
        maxLength: 1000
        type: string
      id:
        type: integer
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      status:
        enum:
        - open
        - in_progress
        - closed
        type: string
      title:
        maxLength: 255
        minLength: 5
        type: string
      updated_at:
        type: string
    required:
    - title
    type: object
...
  models.UpdateCommentRequest:
    properties:
      comment:
        type: string
      ticket_id:
        type: integer
    required:
    - ticket_id
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  types.Pagination:
    properties:
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: API documentation
  title: Example API
  version: "1.0"
paths:
...
  /api/v1/comments:
    get:
      description: Retrieve comments with optional pagination (page, size) and filter
        by user_id
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: size
        type: integer
      - description: Filter by user_id
        in: query
        name: user_id
        type: integer
      - description: Filter by priority
        in: query
        name: priority
        type: string
      - description: Filter by assignee_id
        in: query
        name: assignee_id
        type: integer
      - description: Filter by reporter_id
        in: query
        name: reporter_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetCommentsResponse'
      security:
      - BearerAuth: []
      summary: Get comments with pagination and filters
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create a new comment item
      parameters:
      - description: Comment object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.CreateCommentResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new comment
      tags:
      - comments
  /api/v1/comments/{id}:
    delete:
      description: Delete a specific comment by its ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.DeleteCommentResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Comment not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete comment by ID
      tags:
      - comments
    get:
      description: Fetch a specific comment by its ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetCommentResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Comment not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get comment by ID
      tags:
      - comments
    patch:
      consumes:
      - application/json
      description: Update a comment item by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.UpdateCommentResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "404":
          description: Comment not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update an existing comment
      tags:
      - comments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"